Refactored Gensen local notebook according to latest changes to preprocess utils. Fixed a minor bug in preprocess utils
